requirements.txt
a2wsgi==1.10.8
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiologic==0.14.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-amazon==9.8.0
apache-airflow-providers-apache-kafka==1.9.0
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-http==5.3.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-snowflake==6.3.1
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.2
argcomplete==3.6.2
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.3
avro==1.12.0
backports-datetime-fromisoformat==2.0.3
backports.strenum==1.2.8
beautifulsoup4==4.13.4
blinker==1.9.0
blobfile==3.0.0
boto3==1.38.36
botocore==1.38.36
cachetools==6.2.0
cadwyn==5.4.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorlog==6.9.0
comm==0.2.2
confluent-kafka==2.10.1
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastavro==1.12.0
filelock==3.18.0
finnhub-python==2.4.24
Flask==2.2.5
flatbuffers==25.2.10
fonttools==4.58.1
fredapi==0.5.2
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.72.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
libclang==18.1.1
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.20.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
ml_dtypes==0.5.1
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.4.4
namex==0.1.0
nest-asyncio==1.6.0
newsdataapi==0.1.23
numpy==1.26.4
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
polygon-api-client==1.14.5
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
PyAthena==3.14.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.2
pytest-mock==3.15.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
python3-saml==1.16.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redshift-connector==2.1.7
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retryhttp==1.3.3
rich==14.0.0
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
s3transfer==0.13.0
safetensors==0.5.3
sagemaker_studio==1.0.15
scipy==1.15.3
scramp==1.4.5
seaborn==0.13.2
sentencepiece==0.2.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
snowflake-sqlalchemy==1.7.4
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
text-unidecode==1.3
tf_keras==2.19.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
update-checker==0.18.0
urllib3==2.4.0
uuid6==2025.0.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
watchtower==3.4.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
xmlsec==1.3.14
yarl==1.20.1
zipp==3.23.0
